requirements.txt
# 版本不一定对
fastapi
uvicorn[standard]
tortoise-orm[aiomysql]
aerich
pydantic[email]
python-multipart
bcrypt
PyJWT
redis
python-dotenv   
requests
cos-python-sdk-v5
Pillow